flask/requirements.txt
groq
python-dotenv
flask
flask-cors
PyPDF2